Move interface directory path
